refactor: HttpRequest 객채에 method, path, body 추출 책임 추가
