+++ app/src/main/java/com/codesoom/assignment/network/HttpRouter.java
@@ -2,16 +2,9 @@ package com.codesoom.assignment.network;
 
 import com.sun.net.httpserver.HttpExchange;
 
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.URI;
 import java.util.HashMap;
 import java.util.Iterator;
-import java.util.Optional;
-import java.util.logging.Logger;
-import java.util.stream.Collectors;
 
 /**
  * 등록된 path, method에 맞는 executor를 실행해줍니다
@@ -71,58 +64,24 @@ public class HttpRouter {
      * @throws IOException responseCode, content를 작성할 때 에러가 발생할 수 있습니다
      */
     public void route(HttpExchange exchange) throws IOException {
-        final String method = exchange.getRequestMethod();
+        final HttpRequest request = new HttpRequest(exchange);
         final HttpResponse responder = new HttpResponse(exchange);
 
-        if (method == null) {
-            responder.send(400, null);
-            return;
-        }
-
-        final URI uri = exchange.getRequestURI();
-        final String path = uri.getPath();
-        if (path == null) {
-            responder.send(400, null);
-            return;
-        }
-
-        final HttpMethod methodType = HttpMethod.convert(method);
-        if (methodType == null) {
+        if (request.isNotValid()) {
             responder.send(404, null);
             return;
-        }
-
-        final String body = getRequestBody(exchange).orElse("");
-
-        Logger.getGlobal().info(method + " " + path);
-        if (!body.isEmpty()) {
-            Logger.getGlobal().info(body);
         }
 
         Iterator<HttpRouterKey> keys = pathMap.keySet().iterator();
         while(keys.hasNext()) {
             HttpRouterKey key = keys.next();
 
-            if (key.equalsMethod(methodType) && key.matchesPath(path)) {
+            if (key.matches(request)) {
                 final RouterExecutable executor = pathMap.get(key);
-                final HttpRequest request = new HttpRequest(path, body);
-
                 executor.execute(request, responder);
                 return;
             }
         }
-    }
-
-    private Optional<String> getRequestBody(HttpExchange exchange) {
-        final InputStream inputStream = exchange.getRequestBody();
-        if (inputStream == null) {
-            return Optional.empty();
-        }
-
-        final String body = new BufferedReader(new InputStreamReader(inputStream))
-                .lines()
-                .collect(Collectors.joining("\n"));
-        return Optional.ofNullable(body);
     }
 
 }
